Refactor: move editor camera from renderer to editor
