+++ Ignis/src/Ignis/Editor/EditorCamera.cpp
@@ -1,9 +1,7 @@
 #include "EditorCamera.h"
 #include "Ignis/Core/Input.h"
 #include <glm/gtc/matrix_transform.hpp>
-
-#define GLM_ENABLE_EXPERIMENTAL
-#include <glm/gtx/quaternion.hpp>
+#include <glm/gtc/quaternion.hpp>
 
 namespace ignis
 {
